--- a/aten/src/ATen/cpu/vec256/vec256_float.h
+++ b/aten/src/ATen/cpu/vec256/vec256_float.h
@@ -225,11 +225,10 @@ Vec256<float> inline operator/(const Vec256<float>& a, const Vec256<float>& b) {
   return _mm256_div_ps(a, b);
 }
 
-// Frac. Implement this here so we can use subtraction
+// frac. Implement this here so we can use subtraction
 Vec256<float> Vec256<float>::frac() const {
   return *this - this->trunc();
 }
-
 
 // Implements the IEEE 754 201X `maximum` operation, which propagates NaN if
 // either input is a NaN.