… aten/src/ATen/native/cuda/CUDAUnaryOps.cpp
… aten/src/ATen/native/native_functions.yaml
… test/common_methods_invocations.py