@cpuhrsch 2019-04-08
Did you look at [Declarations.cwrap](https://github.com/pytorch/pytorch/blob/master/aten/src/ATen/Declarations.cwrap) to see if you can remove some legacy functions or restricted them to CUDA etc.?