@cpuhrsch 2019-04-09
Could you revisit the relevant tests for this and check for dtype coverage and large input tensors? This has been an issue in the past.